+++ plugins/plotly-express/src/deephaven/plot/express/deephaven_figure/__init__.py
@@ -3,5 +3,5 @@ from .DeephavenFigure import (
     DeephavenFigureNode,
 )
 from .generate import generate_figure, update_traces
-from .custom_draw import draw_ohlc, draw_candlestick
+from .custom_draw import draw_ohlc, draw_candlestick, draw_density_heatmap
 from .RevisionManager import RevisionManager