+++ plugins/ui/src/deephaven/ui/components/radio_group.py
@@ -1,7 +1,97 @@
+from __future__ import annotations
+from typing import Any, Callable
+from .types import (
+    # Events
+    FocusEventCallable,
+    # Layout
+    AlignSelf,
+    CSSProperties,
+    DimensionValue,
+    JustifySelf,
+    LayoutFlex,
+    Position,
+    LabelPosition,
+    Align,
+    # Validation
+    NecessityIndicator,
+    Orientation,
+    ValidationBehavior,
+)
 from .basic import component_element
+from ..elements import Element
+from .._internal.utils import create_props
 
-# TODO: Add pydocs #514
-def radio_group(*children, **props):
+
+def radio_group(
+    *children,
+    is_emphasized: bool | None = None,
+    orientation: Orientation = "vertical",
+    value: str | None = None,
+    default_value: str | None = None,
+    is_disabled: bool | None = None,
+    is_read_only: bool | None = None,
+    name: str | None = None,
+    is_required: bool | None = None,
+    is_invalid: bool | None = None,
+    validation_behavior: ValidationBehavior | None = None,
+    # validate, # omitted because of synchronouse return
+    label: Any | None = None,
+    description: Any | None = None,
+    error_message: Any | None = None,
+    label_position: LabelPosition = "top",
+    label_align: Align = "start",
+    necessity_indicator: NecessityIndicator = "icon",
+    contextual_help: Any | None = None,
+    show_error_icon: bool | None = None,
+    on_focus: FocusEventCallable | None = None,
+    on_blur: FocusEventCallable | None = None,
+    on_focus_change: Callable[[bool], None] | None = None,
+    on_change: Callable[[str], None] | None = None,
+    flex: LayoutFlex | None = None,
+    flex_grow: float | None = None,
+    flex_shrink: float | None = None,
+    flex_basis: DimensionValue | None = None,
+    align_self: AlignSelf | None = None,
+    justify_self: JustifySelf | None = None,
+    order: int | None = None,
+    grid_area: str | None = None,
+    grid_row: str | None = None,
+    grid_row_start: str | None = None,
+    grid_row_end: str | None = None,
+    grid_column: str | None = None,
+    grid_column_start: str | None = None,
+    grid_column_end: str | None = None,
+    margin: DimensionValue | None = None,
+    margin_top: DimensionValue | None = None,
+    margin_bottom: DimensionValue | None = None,
+    margin_start: DimensionValue | None = None,
+    margin_end: DimensionValue | None = None,
+    margin_x: DimensionValue | None = None,
+    margin_y: DimensionValue | None = None,
+    width: DimensionValue | None = None,
+    height: DimensionValue | None = None,
+    min_width: DimensionValue | None = None,
+    min_height: DimensionValue | None = None,
+    max_width: DimensionValue | None = None,
+    max_height: DimensionValue | None = None,
+    position: Position | None = None,
+    top: DimensionValue | None = None,
+    bottom: DimensionValue | None = None,
+    start: DimensionValue | None = None,
+    end: DimensionValue | None = None,
+    left: DimensionValue | None = None,
+    right: DimensionValue | None = None,
+    z_index: int | None = None,
+    is_hidden: bool | None = None,
+    id: str | None = None,
+    aria_label: str | None = None,
+    aria_labelledby: str | None = None,
+    aria_describedby: str | None = None,
+    aria_details: str | None = None,
+    aria_errormessage: str | None = None,
+    UNSAFE_class_name: str | None = None,
+    UNSAFE_style: CSSProperties | None = None,
+) -> Element:
     """
     Radio buttons allow users to select a single option from a list of mutually
     exclusive options. All possible options are exposed up front for users to
@@ -9,6 +99,75 @@ def radio_group(*children, **props):
 
     Args:
         children: The Radio(s) contained within the RadioGroup.
-        **props: Any other RadioGroup props.
+        is_emphasized: By default, radio button are not emphasized (gray). The emphasized version provides visual prominence.
+        orientation: The axis the Radio Buttons should align with.
+        value: The value of the selected radio button.
+        default_value: The default value of the radio button.
+        is_disabled: Whether the radio button is disabled.
+        is_read_only: Whether the radio button can be selected but not changed by the user.
+        name: The name of the radio button, used when submitting and HTML form.
+        is_required: Whether the radio button is required on the input before form submission.
+        is_invalid: Whether the radio button is in an invalid state.
+        validation_behavior: Whether to use native HTML form validation to prevent form submission when the value is missing or invalid, or mark the field as required or invalid via ARIA.
+        label: The content to display as the label.
+        description: A description for the field. Provides a hint such as specific requirements for what to choose.
+        error_message: An error message for the field.
+        label_position: The position of the label relative to the radio button.
+        label_align: The horizontal alignment of the label relative to the radio button.
+        necessity_indicator: Whether the required state should be shown as an icon or text.
+        contextual_help: A ContextualHelp element to place next to the label.
+        show_error_icon: Whether an error icon is rendered.
+        on_change: Handler that is called when the radio button value changes.
+        on_focus: Handler that is called when the radio button is focused.
+        on_blur: Handler that is called when the radio button loses focus.
+        on_focus_change: Handler that is called when the radio button gains or loses focus.
+        flex: When used in a flex layout, specifies how the element will grow or shrink to fit the space available.
+        flex_grow: When used in a flex layout, specifies how the element will grow to fit the space available.
+        flex_shrink: When used in a flex layout, specifies how the element will shrink to fit the space available.
+        flex_basis: When used in a flex layout, specifies the initial main size of the element.
+        align_self: Overrides the alignItems property of a flex or grid container.
+        justify_self: Species how the element is justified inside a flex or grid container.
+        order: The layout order for the element within a flex or grid container.
+        grid_area: When used in a grid layout specifies, specifies the named grid area that the element should be placed in within the grid.
+        grid_row: When used in a grid layout, specifies the row the element should be placed in within the grid.
+        grid_column: When used in a grid layout, specifies the column the element should be placed in within the grid.
+        grid_row_start: When used in a grid layout, specifies the starting row to span within the grid.
+        grid_row_end: When used in a grid layout, specifies the ending row to span within the grid.
+        grid_column_start: When used in a grid layout, specifies the starting column to span within the grid.
+        grid_column_end: When used in a grid layout, specifies the ending column to span within the grid.
+        margin: The margin for all four sides of the element.
+        margin_top: The margin for the top side of the element.
+        margin_bottom: The margin for the bottom side of the element.
+        margin_start: The margin for the logical start side of the element, depending on layout direction.
+        margin_end: The margin for the logical end side of the element, depending on layout direction.
+        margin_x: The margin for the left and right sides of the element.
+        margin_y: The margin for the top and bottom sides of the element.
+        width: The width of the element.
+        min_width: The minimum width of the element.
+        max_width: The maximum width of the element.
+        height: The height of the element.
+        min_height: The minimum height of the element.
+        max_height: The maximum height of the element.
+        position: The position of the element.
+        top: The distance from the top of the containing element.
+        bottom: The distance from the bottom of the containing element.
+        left: The distance from the left of the containing element.
+        right: The distance from the right of the containing element.
+        start: The distance from the start of the containing element, depending on layout direction.
+        end: The distance from the end of the containing element, depending on layout direction.
+        z_index: The stack order of the element.
+        is_hidden: Whether the element is hidden.
+        id: The unique identifier of the element.
+        aria-label: Defines a string value that labels the current element.
+        aria-labelledby: Identifies the element (or elements) that labels the current element.
+        aria-describedby: Identifies the element (or elements) that describes the object.
+        aria-details: Identifies the element (or elements) that provide a detailed, extended description for the object.
+        aria_errormessage: Identifies the element that provides an error message for the object.
+        UNSAFE_class_name: Set the CSS className for the element. Only use as a last resort. Use style props instead.
+        UNSAFE_style: Set the inline style for the element. Only use as a last resort. Use style props instead.
+
     """
+
+    children, props = create_props(locals())
+
     return component_element(f"RadioGroup", *children, **props)