+++ plugins/ui/src/deephaven/ui/components/types/layout.py
@@ -160,5 +160,6 @@ BorderRadius = Literal[
     "medium",
     "large",
 ]
+MenuDirection = Literal["bottom", "top"]
 IconSize = Literal["XXS", "XS", "S", "M", "L", "XL", "XXL"]
 IconColor = Literal["negative", "notice", "positive", "informative"]